third_party/blink/web_tests/http/tests/websocket/invalid-subprotocol-characters.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("WebSocket should fail if subprotocol contains a forbidden character.");

window.jsTestIsAsync = true;

const URL = "ws://127.0.0.1:8880/accept-first-subprotocol";
const SEPARATORS = "()<>@,;:\\\"/[]?={} \t";

function escapeUnicodeCharacter(codePoint)
{
  let hexCode = codePoint.toString(16);
  while (hexCode.length < 4)
    hexCode = "0" + hexCode;
  return "\\u" + hexCode;
}

function toCharacter(codePoint)
{
  return String.fromCharCode(codePoint);
}

function isValidSubprotocolCodepoint(codePoint)
{
  return codePoint >= 0x21 && codePoint <= 0x7E &&
      SEPARATORS.indexOf(toCharacter(codePoint)) === -1;
}

function assertValidSubprotocolAndFinish(subProtocol)
{
  const ws = new WebSocket(URL, subProtocol);
  ws.onopen = () => {
    testPassed(`WebSocket correctly accepted subprotocol "${subProtocol}"`);
    ws.close();
  };
  ws.onclose = () => finishJSTest();
}

function runTest()
{
  let validSubProtocol = "";
  for (let codePoint = 0; codePoint < 128; ++codePoint) {
    if (isValidSubprotocolCodepoint(codePoint)) {
      // We concatenate all the valid codepoints into a single string so we
      // can test them all at once.
      validSubProtocol += toCharacter(codePoint);
    } else {
      shouldThrow(`new WebSocket(URL, "${escapeUnicodeCharacter(codePoint)}")`);
    }
  }
  assertValidSubprotocolAndFinish(validSubProtocol);
}

runTest();

</script>
</body>
</html>
